Revert "Implemented interval delta"
This reverts commit ec0c552be1c0bb0b6eb9b504b7c34b1338becc1b.

+++ Unity/Assets/Script/Roga2d/Interval/Roga2dAlphaInterval.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Roga2dAlphaInterval : Roga2dValueInterval<float> {
 	private Roga2dNode node;
 	
-	public Roga2dAlphaInterval(Roga2dNode node, float start, float end, float duration, bool tween) 
+	public Roga2dAlphaInterval(Roga2dNode node, float start, float end, int duration, bool tween) 
 	: base(start, end, duration, tween) 
 	{
 		this.node = node;
